add policy to controllers of reshteh and field

+++ app/controllers/fields_controller.rb
@@ -1,10 +1,12 @@
 class FieldsController < ApplicationController
   def index
+    authorize(Prof)
     fields = Field.all
     render jsonapi: fields
   end
 
   def create
+    authorize(Prof)
     field1 = Field.new(field_params)
     if field1.save
         render json: {fieldSaved: true}
@@ -14,13 +16,16 @@ class FieldsController < ApplicationController
   end
 
   def show
+    authorize(Prof)
     render jsonapi: Field.find(params[:id])
   end
 
   def edit
+    authorize(Prof)
   end
 
   def update
+    authorize(Prof)
     field = Field.find(params[:id])
     if field.update(field_params)
       render json: { newtitle: field.title}
@@ -30,6 +35,7 @@ class FieldsController < ApplicationController
   end
 
   def destroy
+    authorize(Prof)
     field = Field.find(params[:id])
     field.destroy
     render json: {fieldcounted: Field.count}
